DDF-5871 removed unused location-new code
