+++ ui/packages/catalog-ui-search/src/main/webapp/component/component.less
@@ -36,7 +36,6 @@
 @import 'input/bulk/input-bulk';
 @import 'input/enum/input-enum';
 @import 'input/radio/input-radio';
-@import 'input/geometry/input-geometry';
 @import 'input/range/input-range';
 @import 'input/autocomplete/input-autocomplete';
 @import 'lightbox/lightbox';
@@ -82,7 +81,6 @@
 @import 'loading-companion/loading-companion';
 @import 'select/select';
 @import 'dropdown/dropdown';
-@import 'location-new/location-new';
 @import 'result-item/result-item';
 @import 'query-feed/query-feed';
 @import 'dropdown/result-display/dropdown.result-display.less';
