--- ui/packages/catalog-ui-search/src/main/webapp/component/input/geometry/input-geometry.less
@@ -1,18 +1,0 @@
-@{customElementNamespace}input.is-geometry {
-  label {
-    white-space: normal;
-    word-break: break-all;
-  }
-}
-
-@{customElementNamespace}input.is-geometry.is-editing {
-  > .if-editing {
-    position: relative;
-    top: unset;
-    transform: unset;
-  }
-
-  > .if-viewing {
-    display: none;
-  }
-}
