--- ui/packages/catalog-ui-search/src/main/webapp/component/location-new/geo-components/geo-components.less
@@ -1,20 +1,0 @@
-.coordinate {
-  display: flex;
-  flex: 1;
-
-  input {
-    flex: 1;
-  }
-
-  @{reactNamespace}direction > .toggle-input {
-    height: @minimumButtonSize;
-    width: @minimumButtonSize;
-    cursor: pointer;
-    text-align: center;
-    padding: 0;
-
-    &:focus {
-      outline: 0;
-    }
-  }
-}
