--- ui/packages/catalog-ui-search/src/main/webapp/component/location-new/inputs/inputs.less
@@ -1,39 +1,0 @@
-@{reactNamespace}text-field {
-  flex: 1;
-
-  input {
-    display: inline-block;
-    width: 100%;
-    white-space: nowrap;
-    height: @minimumButtonSize;
-    line-height: @minimumButtonSize;
-  }
-}
-
-@{reactNamespace}list-editor {
-  .list {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-  }
-
-  .item {
-    margin: 1px;
-  }
-
-  .button-remove {
-    padding: 1px;
-    display: block;
-    width: @minimumButtonSize;
-    height: @minimumButtonSize;
-    font-size: @largeFontSize;
-  }
-
-  .button-add {
-    padding: 1px;
-    display: block;
-    width: 100%;
-    height: @minimumButtonSize;
-    font-size: @largeFontSize;
-  }
-}
