--- ui/packages/catalog-ui-search/src/main/webapp/component/location-new/location-new.less
@@ -1,67 +1,0 @@
-@{customElementNamespace}location-new {
-  display: block;
-  width: auto;
-
-  .form-group {
-    .input-group-addon {
-      border: none;
-      display: table-cell;
-      font-size: @minimumFontSize;
-      padding: 0px @minimumSpacing;
-      height: @minimumButtonSize !important;
-      line-height: @minimumButtonSize !important;
-      width: @minimumButtonSize !important;
-    }
-
-    .input-group {
-      border: none;
-      display: flex;
-      height: @minimumButtonSize;
-      line-height: @minimumButtonSize;
-
-      > span.input-group-addon,
-      > label.input-group-addon {
-        font-weight: lighter;
-        opacity: @minimumOpacity;
-      }
-
-      > span {
-        min-width: 9.375rem;
-        background: inherit;
-        font-size: @minimumFontSize;
-      }
-
-      > input {
-        height: 100%;
-      }
-
-      > label {
-        font-size: @minimumFontSize;
-      }
-    }
-
-    .for-error {
-      display: block;
-      margin-top: @minimumSpacing;
-      padding-right: @minimumSpacing;
-      white-space: pre-line;
-    }
-
-    min-height: 1px;
-    margin-bottom: @minimumSpacing;
-    padding-right: 0px;
-    padding-left: 0px;
-    max-width: 2 * @minimumScreenSize;
-    float: none;
-  }
-  .input-location {
-    border-radius: @border-radius;
-  }
-
-  .input-location + .input-location {
-    margin-top: 10px;
-  }
-}
-
-@import './inputs/inputs.less';
-@import './geo-components/geo-components.less';
