--- ui/packages/catalog-ui-search/src/main/webapp/component/location-new/models/usng-model.js
@@ -1,32 +1,0 @@
-/**
- * Copyright (c) Codice Foundation
- *
- * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser
- * General Public License as published by the Free Software Foundation, either version 3 of the
- * License, or any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
- * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details. A copy of the GNU Lesser General Public License
- * is distributed along with this program and can be found at
- * <http://www.gnu.org/licenses/lgpl.html>.
- *
- **/
-const usngModel = {
-  shape: 'point',
-  point: '',
-  circle: {
-    point: '',
-    radius: '',
-    units: 'meters',
-  },
-  line: {
-    list: [],
-  },
-  polygon: {
-    list: [],
-  },
-  boundingbox: '',
-}
-
-module.exports = usngModel
