+++ ui/packages/catalog-ui-search/src/main/webapp/styles/mixins.less
@@ -268,17 +268,6 @@
       border-right: 1px solid fade(@contrastColor, 40%);
     }
   }
-  @{customElementNamespace}location-new {
-    .input-location {
-      border: 1px solid fade(@contrastColor, 40%);
-    }
-    .input-location > .input-group > .input-group-addon {
-      border-right: 1px solid fade(@contrastColor, 40%);
-    }
-    .input-location input {
-      border: 2px solid fade(@contrastColor, 10%);
-    }
-  }
 }
 
 .is-input();
